"Chagned the floats to doubles"

+++ Utils/Glue_utilities.c
@@ -66,7 +66,7 @@ else{
   for (i = 0; i < (*dim)-1; i++)
   {
 	/*obtain minimum value*/
-	scan_args = sscanf(*ts, " %f , %n", (&((*mins)[i])), &characters_read);
+	scan_args = sscanf(*ts, " %lf , %n", (&((*mins)[i])), &characters_read);
 	if(scan_args == 1)
 	{
 		/*if value is read correctly*/
@@ -75,7 +75,7 @@ else{
 	else if(scan_args ==0)
 	{
 		/*if no value is read (ie we are using negative inf as the min)*/
-		((*mins)[i]) = FLT_MIN;
+		((*mins)[i]) = DBL_MIN;
 		sscanf(*ts, " , %n", &characters_read);
 		*ts = *ts + characters_read;
 	}
@@ -84,7 +84,7 @@ else{
 		exit(0);
 	}
 	/*obtain maximum value*/
-    scan_args = sscanf(*ts," %f ] _ [%n",(&((*maxs)[i])),&characters_read);
+    scan_args = sscanf(*ts," %lf ] _ [%n",(&((*maxs)[i])),&characters_read);
 	if(scan_args == 1)
 	{
 		/*if value is correctly read*/
@@ -92,7 +92,7 @@ else{
 	}
 	else if(scan_args ==0){
 		/*if no value is read (ie we are using positive inf as the max)*/
-		((*maxs)[i]) = FLT_MAX;
+		((*maxs)[i]) = DBL_MAX;
 		sscanf(*ts, " ] _ [%n", &characters_read);
 		*ts = *ts + characters_read;
 	}
@@ -104,7 +104,7 @@ else{
 
   /*getting the last min max value pair*/
 	/*obtain minimum value*/
-	scan_args = sscanf(*ts, " %f , %n",  (&((*mins)[i])), &characters_read);
+	scan_args = sscanf(*ts, " %lf , %n",  (&((*mins)[i])), &characters_read);
 	if(scan_args == 1)
 	{
 		/*if value is read correctly*/
@@ -113,7 +113,7 @@ else{
 	else if(scan_args ==0)
 	{
 		/*if no value is read (ie we are using negative inf as the min)*/
-		((*mins)[i]) = FLT_MIN;
+		((*mins)[i]) = DBL_MIN;
 		sscanf(*ts, " , %n", &characters_read);
 		*ts = *ts + characters_read;
 	}
@@ -122,7 +122,7 @@ else{
 		exit(0);
 	}
 	/*obtain maximum value*/
-    scan_args = sscanf(*ts," %f ] :%n",&((*maxs)[i]),&characters_read);
+    scan_args = sscanf(*ts," %lf ] :%n",&((*maxs)[i]),&characters_read);
 	if(scan_args == 1)
 	{
 		/*if value is correctly read*/
@@ -130,7 +130,7 @@ else{
 	}
 	else if(scan_args ==0){
 		/*if no value is read (ie we are using positive inf as the max)*/
-		(*maxs)[i] = FLT_MAX;
+		(*maxs)[i] = DBL_MAX;
 		sscanf(*ts, " ] :%n", &characters_read);
 		  *ts = *ts + characters_read;
 	}
